refactor: move MMO logic into StepProjectDetails
Moved isMMO() helper and userCanViewDeleteButton logic from ProjectSettings
into StepProjectDetails component where it's actually used.

Benefits:
- Reduces prop drilling (one less prop to pass)
- Logic is closer to where it's used
- StepProjectDetails is more self-contained
- Follows React best practices
